deployment/requirements.txt
flask==3.0.2
tensorflow-cpu==2.16.1
numpy==1.26.4
scikit-learn==1.4.0
pyngrok==7.0.0